+++ budget-manager-server/src/BudgetManager.Infrastructure/Data/Configurations/UserConfiguration.cs
@@ -2,7 +2,6 @@
 using BudgetManager.Domain.Models.ValueObjects;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using System.Reflection.Emit;
 
 namespace BudgetManager.Infrastructure.Data.Configurations;
